requirements.txt
Flask>=2.3
python-dotenv>=1.0